did header and navbar
