+++ src/components/Navbar/Navbar.module.css
@@ -2,10 +2,8 @@
   display: flex;
   flex-direction: column;
   background-color: #17252a;
-  min-height: 92vh;
+  min-height: 71vh;
   max-height: 100%;
-  width: 16vw;
-  min-width: 150px;
 }
 
 .item:last-child {
